Add files via upload
Python files created for the Capstone projects for the IBM Professional Certificate in Data Science. The project describes a case study of analysing the launch data for rocket missions run by SpaceX to understand the characteristics of successful launches, with the development of a Dash dashboard (Module 3.2), a Folium map (Module 3.1) and a predictive model (Module 4).